+++ src/tools/toolkit.py
@@ -1,21 +1,11 @@
 """this module contains high level utility functions that makes it easier to use tools by providing wrappers over tools"""
 
-from typing import Literal
-
 from beartype import beartype
 from deeporigin.data_hub import api
-from deeporigin.exceptions import DeepOriginException
 from deeporigin.tools import run
-from deeporigin.utils.config import construct_resource_url
 
 VINA_DB = "autodock-vina"
 LIGAND_PREP_DB = "ligand-prep-meeko"
-ABFE_DB = "ABFE"
-
-
-charge_methods = Literal["gas", "bcc"]
-lig_force_fields = Literal["gaff2", "openff"]
-force_fields = Literal["ff14SB", "ff99SB-ildn"]
 
 
 @beartype
@@ -28,7 +18,7 @@ def _ensure_database(name: str) -> dict:
 
     if len(database) == 0:
         # make a new DB
-        print("🧬 Creating a database called {name}...")
+        print(f"🧬 Creating a database called {name}...")
         database = api.create_database(
             hid=name,
             hid_prefix=name,
@@ -67,137 +57,6 @@ def _ensure_columns(
             required=False,
             type=column_type,
         )
-
-
-@beartype
-def _ensure_db_for_abfe() -> dict:
-    """ensure that there is a database for FEP on Data Hub"""
-
-    required_columns = [
-        dict(name="ligand", type="file"),
-        dict(name="protein", type="file"),
-        dict(name="prep_output", type="file"),
-        dict(name="emeq_output", type="file"),
-    ]
-
-    database = _ensure_database(ABFE_DB)
-
-    _ensure_columns(
-        database=database,
-        required_columns=required_columns,
-    )
-
-    return database
-
-
-def system_prep(
-    *,
-    ligand_file: str,
-    protein_file: str,
-    keep_waters: bool = True,
-    save_gmx_files: bool = False,
-    is_lig_protonated: bool = True,
-    is_protein_protonated: bool = True,
-    do_loop_modelling: bool = False,
-    charge_method: charge_methods = "bcc",
-    lig_force_field: lig_force_fields = "gaff2",
-    padding: float = 1.0,
-    system_name: str = "complex",
-    force_field: force_fields = "ff14SB",
-) -> None:
-    """High level function to prepare Ligand and protein files using Deep Origin MDSuite. Use this function to run system prep on a ligand and protein pair, that exist as files on your local computer.
-
-    Args:
-        ligand_file (str): path to ligand file
-        protein_file (str): path to protein file
-        keep_waters (bool, optional): whether to keep waters. Defaults to True.
-        save_gmx_files (bool, optional): whether to save gmx files. Defaults to False.
-        is_lig_protonated (bool, optional): whether the ligand is protonated. Defaults to True.
-        is_protein_protonated (bool, optional): whether the protein is protonated. Defaults to True.
-        do_loop_modelling (bool, optional): whether to do loop modelling. Defaults to False.
-        charge_method (charge_methods, optional): charge method. Defaults to "bcc".
-        lig_force_field (lig_force_fields, optional): ligand force field. Defaults to "gaff2".
-        padding (float, optional): padding. Defaults to 1.0.
-        system_name (str, optional): system name. Defaults to "complex".
-        force_field (force_fields, optional): force field. Defaults to "ff14SB".
-
-    """
-
-    # input validation
-    if padding < 0.5 or padding > 2:
-        raise DeepOriginException("Padding must be greater than 0.5, and less than 2")
-
-    database = _ensure_db_for_abfe()
-
-    url = construct_resource_url(
-        name=ABFE_DB,
-        row_type="database",
-    )
-
-    print(f"Using database at: {url}")
-    print("🧬 Uploading files to database...")
-
-    response = api.upload_file_to_new_database_row(
-        database_id=ABFE_DB,
-        column_id="ligand",
-        file_path=ligand_file,
-    )
-
-    row_id = response.rows[0].hid
-
-    file = api.upload_file(file_path=protein_file)
-
-    api.assign_files_to_cell(
-        file_ids=[file.id],
-        database_id=ABFE_DB,
-        column_id="protein",
-        row_id=row_id,
-    )
-
-    tool_id = "deeporigin/md-suite-prep"
-    inputs = {
-        "ligand": {
-            "columnId": "ligand",
-            "rowId": row_id,
-            "databaseId": database.hid,
-        },
-        "protein": {
-            "columnId": "protein",
-            "rowId": row_id,
-            "databaseId": database.hid,
-        },
-        "force": 2,
-        "test_run": 0,
-        "system": system_name,
-        "include_ligands": 1,
-        "include_protein": 1,
-        "sysprep_params": {
-            "is_protein_protonated": is_protein_protonated,
-            "do_loop_modelling": do_loop_modelling,
-            "force_field": "ff14SB",
-            "padding": padding,
-            "keep_waters": keep_waters,
-            "save_gmx_files": save_gmx_files,
-            "is_lig_protonated": is_lig_protonated,
-            "charge_method": charge_method,
-            "lig_force_field": lig_force_field,
-        },
-    }
-
-    outputs = {
-        "output_file": {
-            "columnId": "prep_output",
-            "rowId": row_id,
-            "databaseId": database.hid,
-        }
-    }
-
-    run._process_job(
-        inputs=inputs,
-        outputs=outputs,
-        tool_id=tool_id,
-        cols=database.cols,
-    )
 
 
 @beartype
